+++ src/max_div/_core/solver/_solver_config.py
@@ -11,9 +11,9 @@ from dataclasses import dataclass, replace
 from max_div._core.constraints import Constraint
 from max_div._core.metrics import DiversityMetric, DiversityObjective
 from max_div._core.metrics._distance import DistanceStore
-from max_div._core.solver._distance_storage import DistanceStorageTypes
 
 from ._constraint_penalty import ConstraintPenalty
+from ._distance_storage import DistanceStorageTypes
 from ._duration import E2eBudget
 from ._solver import MaxDivSolver
 from ._solver_step import REPORTING_BATCH_SECONDS, SolverStep